feat: create add-missing-jsx-attribute transform, getParsedImports util
fixup getParsedImports

fixup getParsedImports - return path of `Declaration`, not `Declarator`

fix: in getParsedImports, return both `path` and `declarationPath` for require's

...because the actual path that has the declaration properties is the
Declarator, not Declaration.

There's always only 1 Declaration. So we're adding it just in case,
but it'll probably be uses less often,
because the Declarator is the same structure as the ImportDeclaration

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>